Move dump to IRDumper class
