+++ query/plan/ir/IRDumper.cpp
@@ -1,0 +1,36 @@
+#include "IRDumper.h"
+
+#include <ostream>
+#include <string>
+#include <unordered_map>
+
+#include "EntityPattern.h"
+#include "VariableDependencyGraph.h"
+#include "decl/VarDecl.h"
+
+using namespace db;
+
+void IRDumper::dumpMermaid(const VariableDependencyGraph& graph, std::ostream& out) {
+    out << "flowchart TD\n";
+
+    std::unordered_map<const VariableDependency*, size_t> nodeIds;
+    for (size_t i = 0; i < graph.vars().size(); i++) {
+        nodeIds[&graph.vars()[i]] = i;
+    }
+
+    const auto nodeDef = [&](const VariableDependency* var) {
+        const std::string_view name = var->entity()->getDecl()->getName();
+        const std::string_view label = name.empty() ? "<unnamed>" : name;
+        return "v" + std::to_string(nodeIds.at(var)) + "[\"" + std::string(label) + "\"]";
+    };
+
+    for (const VariableDependency& var : graph.vars()) {
+        if (var.isIsolated()) {
+            out << "    " << nodeDef(&var) << "\n";
+        } else {
+            for (const VariableDependency* dep : var.getOutgoing()) {
+                out << "    " << nodeDef(&var) << " --> " << nodeDef(dep) << "\n";
+            }
+        }
+    }
+}
