+++ query/plan/ir/IRDumper.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <iosfwd>
+
+namespace db {
+
+class VariableDependencyGraph;
+
+class IRDumper {
+public:
+    static void dumpMermaid(const VariableDependencyGraph& graph, std::ostream& out);
+};
+
+}
